+++ src/civis/polling.py
@@ -5,7 +5,7 @@ from civis.base import CivisJobFailure, CivisAsyncResultBase, FAILED, DONE
 from civis.response import Response
 
 
-_DEFAULT_POLLING_INTERVAL = 15
+_MAX_POLLING_INTERVAL = 15
 
 
 class _ResultPollingThread(threading.Thread):
@@ -13,11 +13,9 @@ class _ResultPollingThread(threading.Thread):
 
     # Inspired by `threading.Timer`
 
-    def __init__(self, poller, poller_args, polling_interval):
+    def __init__(self, pollable_result):
         super().__init__(daemon=True)
-        self.polling_interval = polling_interval
-        self.poller = poller
-        self.poller_args = poller_args
+        self.pollable_result = pollable_result
         self.finished = threading.Event()
 
     def cancel(self):
@@ -31,11 +29,11 @@ class _ResultPollingThread(threading.Thread):
 
     def run(self):
         """Poll until done."""
-        while not self.finished.wait(self.polling_interval):
+        while not self.finished.wait(self.pollable_result._next_polling_interval):
             # Spotty internet connectivity can result in polling functions
             # returning None. This treats None responses like responses which
             # have a non-DONE state.
-            poller_result = self.poller(*self.poller_args)
+            poller_result = self.pollable_result._check_result()
             if poller_result is not None and poller_result.state in DONE:
                 self.finished.set()
 
@@ -53,9 +51,14 @@ class PollableResult(CivisAsyncResultBase):
         A function which returns an object that has a ``state`` attribute.
     poller_args : tuple
         The arguments with which to call the poller function.
-    polling_interval : int or float
+    polling_interval : int or float, optional
         The number of seconds between API requests to check whether a result
-        is ready.
+        is ready. If an integer or float is provided, this number will be used
+        as the polling interval. If ``None`` (the default), the polling interval will
+        start at 1 second and increase geometrically up to 15 seconds. The ratio of
+        the increase is 1.2, resulting in polling intervals in seconds of
+        1, 1.2, 1.44, 1.728, etc. This default behavior allows for a faster return for
+        a short-running job and a capped polling interval for longer-running jobs.
     client : :class:`civis.APIClient`, optional
         If not provided, an :class:`civis.APIClient` object will be
         created from the :envvar:`CIVIS_API_KEY`.
@@ -103,17 +106,19 @@ class PollableResult(CivisAsyncResultBase):
         client=None,
         poll_on_creation=True,
     ):
-        if polling_interval is None:
-            polling_interval = _DEFAULT_POLLING_INTERVAL
-        super().__init__(
-            poller=poller,
-            poller_args=poller_args,
-            polling_interval=polling_interval,
-            client=client,
-            poll_on_creation=poll_on_creation,
-        )
-        if self.polling_interval <= 0:
+        super().__init__()
+
+        self.poller = poller
+        self.poller_args = poller_args
+        self.polling_interval = polling_interval
+        self.client = client
+        self.poll_on_creation = poll_on_creation
+
+        if self.polling_interval is not None and self.polling_interval <= 0:
             raise ValueError("The polling interval must be positive.")
+
+        self._next_polling_interval = 1
+        self._use_geometric_polling = True
 
         # Polling arguments. Never poll more often than the requested interval.
         if poll_on_creation:
@@ -153,8 +158,20 @@ class PollableResult(CivisAsyncResultBase):
             now = time.time()
             if (
                 not self._last_polled
-                or (now - self._last_polled) >= self.polling_interval
+                or (now - self._last_polled) >= self._next_polling_interval
             ):
+                if self._use_geometric_polling:
+                    # Choosing a common ratio of 1.2 for these polling intervals:
+                    #   1, 1.2, 1.44, 1.73, 2.07, 2.49, 2.99, ..., and capped at 15.
+                    # Within the first 15 secs by wall time, we call the poller 7 times,
+                    # which gives a short-running job's future.result()
+                    # a higher chance to return faster.
+                    # For longer running jobs, the polling interval will be capped
+                    # at 15 secs when by wall time 87 secs have passed.
+                    self._next_polling_interval *= 1.2
+                    if self._next_polling_interval > _MAX_POLLING_INTERVAL:
+                        self._next_polling_interval = _MAX_POLLING_INTERVAL
+                        self._use_geometric_polling = False
                 # Poll for a new result
                 self._last_polled = now
                 try:
@@ -204,9 +221,7 @@ class PollableResult(CivisAsyncResultBase):
             if self._polling_thread.is_alive():
                 self._polling_thread.cancel()
 
-    def _reset_polling_thread(
-        self, polling_interval=_DEFAULT_POLLING_INTERVAL, start_thread=False
-    ):
+    def _reset_polling_thread(self, polling_interval, start_thread=False):
         with self._condition:
             if (
                 getattr(self, "_polling_thread", None) is not None
@@ -214,8 +229,8 @@ class PollableResult(CivisAsyncResultBase):
             ):
                 self._polling_thread.cancel()
             self.polling_interval = polling_interval
-            self._polling_thread = _ResultPollingThread(
-                self._check_result, (), polling_interval
-            )
+            self._next_polling_interval = 1 if (pi := polling_interval) is None else pi
+            self._use_geometric_polling = polling_interval is None
+            self._polling_thread = _ResultPollingThread(self)
             if start_thread:
                 self._polling_thread.start()